Done with delete and update

+++ src/main/java/uk/ac/addressmanger/dataservice/AddressDao.java
@@ -14,6 +14,6 @@ public interface  AddressDao extends CrudRepository<Address, Long> {
 	   * 
 	   * @param email the user email.
 	   */
-	  public Address findById(int id);
+	  public Address findById(long id);
 
 } // class UserDao
